requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
beautifulsoup4~=4.13.4
numpy~=2.3.2
pandas~=2.3.1
requests~=2.32.4
azure-core~=1.35.0